Fixed 3.4. integration test.

+++ src/main/kotlin/com/senacor/cap/service/message/ChatMessageRepository.kt
@@ -3,7 +3,7 @@ package com.senacor.cap.service.message
 import org.springframework.data.repository.CrudRepository
 
 @org.springframework.stereotype.Repository
-interface ChatMessageRepository : CrudRepository<ChatMessage,Int> {
+interface ChatMessageRepository : CrudRepository<ChatMessage,Long> {
 
     //fun findByChannelId(channelId:String):List<ChatMessage>;
     fun findByChannelIdOrderByCreationTimestampDesc(s: String): List<ChatMessage>
